Base API stubs for admin API
:package: Moved admin controllers into own package
