+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/controller/admin/OpenehrAdminEhrController.java
@@ -15,41 +15,43 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehrbase.rest.openehr.controller;
+package org.ehrbase.rest.openehr.controller.admin;
 
 import io.swagger.annotations.*;
 import org.ehrbase.api.exception.ObjectNotFoundException;
 import org.ehrbase.api.service.EhrService;
 import org.ehrbase.response.openehr.AdminDeleteResponseData;
 import org.ehrbase.response.openehr.AdminUpdateResponseData;
+import org.ehrbase.rest.openehr.controller.BaseController;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
 
+
 /**
- * Admin API controller for Contribution related data. Provides endpoints to update and remove Contributions in
- * database physically.
+ * Admin API controller for EHR related endpoints. Provides methods to update and delete EHRs physically in the DB.
  */
-@Api(tags = {"Admin", "Contribution"})
+@Api(tags = {"Admin", "EHR"})
 @RestController
-@RequestMapping(path = "/rest/openehr/v1/admin/ehr")
-public class OpenehrAdminContributionController extends BaseController {
+@RequestMapping(path = "/rest/openehr/v1/admin/ehr", produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
+public class OpenehrAdminEhrController extends BaseController {
 
     private final EhrService ehrService;
 
     @Autowired
-    public OpenehrAdminContributionController(EhrService ehrService) {
+    public OpenehrAdminEhrController(EhrService ehrService) {
         this.ehrService = ehrService;
     }
 
-    @PutMapping(path = "/{ehr_id}/contribution/{contribution_id}")
+    @PutMapping(path = "/{ehr_id}", consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
     @ApiResponses(value = {
             @ApiResponse(
                     code = 200,
-                    message = "Contribution has been updated successfully.",
+                    message = "EHR has been updated and number of updated items will be returned in the body.",
                     responseHeaders = {
                             @ResponseHeader(
                                     name = CONTENT_TYPE,
@@ -60,44 +62,40 @@ public class OpenehrAdminContributionController extends BaseController {
             ),
             @ApiResponse(
                     code = 401,
-                    message = "Client credentials invalid or have expired."
+                    message = "Client credentials are invalid or have been expired."
             ),
             @ApiResponse(
                     code = 403,
-                    message = "Client does not have permission to access since admin role is missing."
+                    message = "Client is not permitted to access this resource since the admin role is missing."
             ),
             @ApiResponse(
                     code = 404,
-                    message = "EHR or Contribution could not be found."
+                    message = "EHR with id could not be found."
             )
     })
-    public ResponseEntity<AdminUpdateResponseData> updateContribution(
-            @ApiParam(value = "Target EHR id to update contribution inside.", required = true)
+    public ResponseEntity<AdminUpdateResponseData> updateEhr(
+            @ApiParam(value = "Client requested response content type")
+            @RequestHeader(value = HttpHeaders.ACCEPT, required = false)
+                    String accept,
+            @ApiParam(value = "Target EHR id to update", required = true)
             @PathVariable(value = "ehr_id")
-                    String ehrId,
-            @ApiParam(value = "Target Contribution id to update", required = true)
-            @PathVariable(value = "contribution_id")
-                    String contributionId
+                    String ehrId
     ) {
+
+        // Check if EHR with id exists
         UUID ehrUuid = UUID.fromString(ehrId);
-
-        // Check if EHR exists
-        if (!this.ehrService.hasEhr(ehrUuid)) {
-            throw new ObjectNotFoundException(
-                    "Admin Contribution", String.format("EHR with id %s does not exist", ehrId)
-            );
+        if (!ehrService.hasEhr(ehrUuid)) {
+            throw new ObjectNotFoundException("Admin EHR", String.format("EHR with id %s does not exist.", ehrId));
         }
-
-        // Contribution existence check will be done in services
 
         return ResponseEntity.ok().body(new AdminUpdateResponseData(0));
     }
 
-    @DeleteMapping(path = "/{ehr_id}/contribution/{contribution_id}")
+    @DeleteMapping(path = "/{ehr_id}")
     @ApiResponses(value = {
             @ApiResponse(
                     code = 200,
-                    message = "Contribution has been deleted successfully.",
+                    message = "EHR has been deleted successfully",
                     responseHeaders = {
                             @ResponseHeader(
                                     name = CONTENT_TYPE,
@@ -108,35 +106,28 @@ public class OpenehrAdminContributionController extends BaseController {
             ),
             @ApiResponse(
                     code = 401,
-                    message = "Client credentials invalid or have expired."
+                    message = "Client credentials are invalid or have been expired."
             ),
             @ApiResponse(
                     code = 403,
-                    message = "Client does not have permission to access since admin role is missing."
+                    message = "Client is not permitted to access this resource since the admin role is missing."
             ),
             @ApiResponse(
                     code = 404,
-                    message = "EHR or Contribution could not be found."
+                    message = "EHR with id could not be found."
             )
     })
-    public ResponseEntity<AdminDeleteResponseData> deleteContribution(
-            @ApiParam(value = "Target EHR id to update contribution inside.", required = true)
+    public ResponseEntity<AdminDeleteResponseData> deleteEhr(
+            @ApiParam(value = "Target EHR id to delete", required = true)
             @PathVariable(value = "ehr_id")
-                    String ehrId,
-            @ApiParam(value = "Target Contribution id to update", required = true)
-            @PathVariable(value = "contribution_id")
-                    String contributionId
+                    String ehrId
     ) {
+
         UUID ehrUuid = UUID.fromString(ehrId);
-
-        // Check if EHR exists
-        if (!this.ehrService.hasEhr(ehrUuid)) {
-            throw new ObjectNotFoundException(
-                    "Admin Contribution", String.format("EHR with id %s does not exist", ehrId)
-            );
+        // Check if EHR with id exists
+        if (!ehrService.hasEhr(ehrUuid)) {
+            throw new ObjectNotFoundException("Admin EHR", String.format("EHR with id %s does not exist.", ehrId));
         }
-
-        // Contribution existence check will be done in services
 
         return ResponseEntity.ok().body(new AdminDeleteResponseData(0));
     }
